@@ -2,7 +2,9 @@
   "name": "imbo/imbo-mongodb-adapters",
   "type": "library",
   "description": "MongoDB adapters for Imbo",
-  "keywords": [ "imbo" ],
+  "keywords": [
+    "imbo"
+  ],
   "homepage": "https://imbo.io/",
   "license": "MIT",
   "authors": [
@@ -10,23 +12,28 @@
       "name": "Christer Edvartsen",
       "email": "cogo@starzinger.net",
       "homepage": "https://github.com/christeredvartsen"
-    }, {
+    },
+    {
       "name": "Contributors",
       "homepage": "https://github.com/imbo"
     }
   ],
   "require": {
-    "php": ">=8.1",
-    "ext-mongodb": "^1.10",
-    "mongodb/mongodb": "^1.6"
+    "php": ">=8.3",
+    "ext-mongodb": "^1.21",
+    "mongodb/mongodb": "^1.21"
   },
   "require-dev": {
+    "friendsofphp/php-cs-fixer": "^3.75",
     "imbo/imbo": "dev-main",
     "imbo/imbo-adapters-sdk": "dev-main",
     "imbo/imbo-coding-standard": "^2.0",
-    "phpunit/phpunit": "^10.0",
-    "psalm/plugin-phpunit": "^0.18.4",
-    "vimeo/psalm": "^5.5"
+    "phpstan/extension-installer": "^1.4",
+    "phpstan/phpstan": "^2.1",
+    "phpstan/phpstan-deprecation-rules": "^2.0",
+    "phpstan/phpstan-phpunit": "^2.0",
+    "phpunit/phpunit": "^12.0",
+    "symfony/var-dumper": "^7.2"
   },
   "autoload": {
     "psr-4": {
@@ -40,9 +47,9 @@
   },
   "scripts": {
     "test": "vendor/bin/phpunit",
-    "test:coverage": "vendor/bin/phpunit --coverage-html build/coverage --coverage-text",
-    "sa": "vendor/bin/psalm",
-    "cs": "php-cs-fixer fix --dry-run --diff",
+    "test:coverage": "vendor/bin/phpunit --coverage-html build/coverage",
+    "sa": "vendor/bin/phpstan",
+    "cs": "vendor/bin/php-cs-fixer fix --dry-run --diff",
     "ci": [
       "@sa",
       "@test",
@@ -50,6 +57,9 @@
     ]
   },
   "config": {
-    "sort-packages": true
+    "sort-packages": true,
+    "allow-plugins": {
+      "phpstan/extension-installer": true
+    }
   }
 }
