@@ -2,11 +2,12 @@
 namespace Imbo\Database;
 
 use MongoDB\Client;
+use MongoDB\Driver\Exception\RuntimeException;
+use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\Group;
 
-/**
- * @coversDefaultClass Imbo\Database\MongoDB
- * @group integration
- */
+#[CoversClass(MongoDB::class)]
+#[Group('integration')]
 class MongoDBIntegrationTest extends DatabaseTests
 {
     private string $databaseName = 'imbo-database-mongodb-integration-test';
@@ -34,6 +35,13 @@ class MongoDBIntegrationTest extends DatabaseTests
 
         $uri = (string) getenv('MONGODB_URI');
         $client = new Client($uri, $uriOptions);
+
+        try {
+            $client->getDatabase($this->databaseName)->command(['ping' => 1]);
+        } catch (RuntimeException) {
+            $this->markTestSkipped('MongoDB is not running, start it with `docker compose up -d`', );
+        }
+
         $client->dropDatabase($this->databaseName);
         $client->selectCollection($this->databaseName, MongoDB::IMAGE_COLLECTION)->createIndex([
             'user'            => 1,
