@@ -2,11 +2,12 @@
 namespace Imbo\Storage;
 
 use MongoDB\Client;
+use MongoDB\Driver\Exception\RuntimeException;
+use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\Group;
 
-/**
- * @coversDefaultClass Imbo\Storage\GridFS
- * @group integration
- */
+#[CoversClass(GridFS::class)]
+#[Group('integration')]
 class GridFSIntegrationTest extends StorageTests
 {
     private string $databaseName = 'imbo-mongodb-adapters-integration-test';
@@ -34,6 +35,13 @@ class GridFSIntegrationTest extends StorageTests
 
         $uri = (string) getenv('MONGODB_URI');
         $client = new Client($uri, $uriOptions);
+
+        try {
+            $client->getDatabase($this->databaseName)->command(['ping' => 1]);
+        } catch (RuntimeException) {
+            $this->markTestSkipped('MongoDB is not running, start it with `docker compose up -d`', );
+        }
+
         $client->dropDatabase($this->databaseName);
     }
 }
